Feat : add hello-mvc that was mapped HelloController and hello-template.html
If you enter "url:8080/hello-mvc?name=hyungjun" into the address bar,
The web server receives it, and the built-in Tomcat server receives it first.
The built-in Tomcat server notifies spring that hello-mvc has arrived.
In spring, hello-mvc was mapped to HelloController,
Executes the function to which HelloController's hello-mvc is mapped.
By code, the value entered as name is added to the model and hello-template is returned to spring.
Then viewResolver works, finds hello-template received as return and requests processing from thymeleaf template engine.
Then, the template engine returns the converted html to the web browser after conversion.